requirements.txt
smolagents[litellm]>=1.20.0
litellm>=1.0.0
openai>=1.0.0
gradio>=4.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
feedparser>=6.0.0
arxiv>=1.4.0
PyGithub>=1.58.0
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
schedule>=1.2.0
aiohttp>=3.8.0
lxml>=4.9.0
newspaper3k>=0.2.8
textstat>=0.7.0
jieba>=0.42.1
